+++ all-is-cubes/src/math/coord.rs
@@ -28,10 +28,44 @@ pub fn cube_to_midpoint(cube: GridPoint) -> Point3<FreeCoordinate> {
     cube.map(|component| FreeCoordinate::from(component) + 0.5)
 }
 
+/// Convert a point in space to the unit cube that encloses it.
+///
+/// Such cubes are defined to be half-open intervals on each axis; that is,
+/// an integer coordinate is counted as part of the cube extending positively
+/// from that coordinate.
+///
+/// ```
+/// use all_is_cubes::cgmath::Point3;
+/// use all_is_cubes::math::point_to_enclosing_cube;
+///
+/// assert_eq!(point_to_enclosing_cube(Point3::new(1.0, 1.5, -2.5)), Point3::new(1, 1, -3));
+/// ```
+#[inline(always)] // trivial arithmetic that almost certainly should be inlined
+pub fn point_to_enclosing_cube(point: Point3<FreeCoordinate>) -> GridPoint {
+    point.map(|component| component.floor() as GridCoordinate)
+}
+
 /// Compute the squared magnitude of a [`GridVector`].
 ///
 /// [`cgmath::InnerSpace::magnitude2`] would do the same but only for floats.
 #[inline(always)] // trivial arithmetic that almost certainly should be inlined
 pub(crate) fn int_magnitude_squared(v: GridVector) -> GridCoordinate {
     v.x * v.x + v.y * v.y + v.z * v.z
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn point_to_enclosing_cube_inf_nan() {
+        assert_eq!(
+            point_to_enclosing_cube(Point3::new(
+                FreeCoordinate::INFINITY,
+                -FreeCoordinate::INFINITY,
+                FreeCoordinate::NAN
+            )),
+            Point3::new(GridCoordinate::MAX, GridCoordinate::MIN, 0)
+        );
+    }
 }